+++ algorithms/0108-convert-sorted-array-to-binary-search-tree.go
@@ -1,0 +1,14 @@
+package algorithms
+
+func sortedArrayToBST(nums []int) *TreeNode {
+	if len(nums) == 0 {
+		return nil
+	}
+
+	i := len(nums) / 2
+	root := &TreeNode{Val: nums[i]}
+	root.Left = sortedArrayToBST(nums[:i])
+	root.Right = sortedArrayToBST(nums[i+1:])
+
+	return root
+}