+++ CHANGELOG.md
@@ -509,6 +509,10 @@ $$Pow({x}, {n})=Pow\left({x}^2, \dfrac{n}{2}\right) * Pow\left(x, n \mod 2\right
 
 - Recursion
 
+## 0108 - Convert Sorted Array to Binary Search Tree
+
+- Recursion
+
 ## 0415 - Add strings
 
 - Full adder